Move common test to test_runtime
